First functional separated components
App is now just the container for the other components and has the main state variables.
SpotifyApiHandler is a renderless component that handles all the fetch requests.
ThreeScene is the component that has the rendering components.Based on the track_analysis I made some pulse functions that return true when any of the rhythm components of the analysis change. I also added a tempoPulse based on the tempo of each section.